backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Email (Brevo)
sib-api-v3-sdk==7.6.0

# Document Parsing
python-docx==1.1.0
openpyxl==3.1.2

# AWS S3
boto3==1.34.0
python-slugify==8.0.1
thefuzz==0.22.1

# AI / Gemini
google-generativeai==0.8.0

# Validation
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# CORS
starlette==0.35.1

# Utils
python-dotenv==1.0.0
pytz==2025.2